refactor: Move RequestFactoryNoCsrf helper to common/
