+++ lms/djangoapps/courseware/tests/factories.py
@@ -5,7 +5,6 @@ import json
 from functools import partial
 
 import factory
-from django.test.client import RequestFactory
 from factory.django import DjangoModelFactory
 from opaque_keys.edx.keys import CourseKey
 from opaque_keys.edx.locator import CourseLocator
@@ -68,13 +67,3 @@ class StudentInfoFactory(DjangoModelFactory):  # lint-amnesty, pylint: disable=m
     field_name = 'existing_field'
     value = json.dumps('old_value')
     student = factory.SubFactory(UserFactory)
-
-
-class RequestFactoryNoCsrf(RequestFactory):
-    """
-    RequestFactory, which disables csrf checks.
-    """
-    def request(self, **kwargs):
-        request = super().request(**kwargs)
-        setattr(request, '_dont_enforce_csrf_checks', True)  # pylint: disable=literal-used-as-attribute
-        return request
